refactor frontend
